make oneOf with select the default one
