+++ example/src/examples/additionalProperties/data.json
@@ -1,0 +1,1 @@
+{ "test": "data" }
