+++ example/src/examples/additionalProperties/index.ts
