+++ example/src/examples/additionalProperties/schema.json
@@ -1,0 +1,13 @@
+{
+  "$schema": "http://json-schema.org/draft-07/schema#",
+  "type": "object",
+  "properties": {
+    "test": {
+      "type": "string"
+    }
+  },
+  "additionalProperties": {
+    "type": "string",
+    "title": "labels"
+  }
+}
