+++ example/src/examples/additionalProperties/uischema.json
@@ -1,0 +1,4 @@
+{
+  "type": "Control",
+  "scope": "#/"
+}
