+++ example/src/examples/oneOf-tab/uischema.json
@@ -6,7 +6,7 @@
       "label": "Basic Information",
       "scope": "#/properties/addressOrUser",
       "options": {
-        "variant": "select"
+        "variant": "tab"
       }
     }
   ]
