+++ vue2-vuetify/src/complex/OneOfRenderer.vue
@@ -6,40 +6,47 @@
       :path="path"
     />
 
-    <v-tabs v-model="tabIndex">
-      <v-tab
-        @change="handleTabChange"
-        v-for="(oneOfRenderInfo, oneOfIndex) in oneOfRenderInfos"
-        :key="`${control.path}-${oneOfIndex}`"
-      >
-        {{ oneOfRenderInfo.label }}
-      </v-tab>
-    </v-tabs>
-
-    <v-tabs-items v-model="selectedIndex">
-      <v-tab-item
-        v-for="(oneOfRenderInfo, oneOfIndex) in oneOfRenderInfos"
-        :key="`${control.path}-${oneOfIndex}`"
-      >
-        <dispatch-renderer
-          v-if="selectedIndex === oneOfIndex"
-          :schema="oneOfRenderInfo.schema"
-          :uischema="oneOfRenderInfo.uischema"
-          :path="control.path"
-          :renderers="control.renderers"
-          :cells="control.cells"
-          :enabled="control.enabled"
-        />
-      </v-tab-item>
-    </v-tabs-items>
+    <v-hover v-slot="{ hover }">
+      <v-select
+        v-disabled-icon-focus
+        :id="control.id + '-input'"
+        :class="styles.control.input"
+        :disabled="!control.enabled"
+        :autofocus="appliedOptions.focus"
+        :placeholder="appliedOptions.placeholder"
+        :label="computedLabel"
+        :hint="control.description"
+        :persistent-hint="persistentHint()"
+        :required="control.required"
+        :error-messages="control.errors"
+        :clearable="hover"
+        :items="indexedOneOfRenderInfos"
+        @change="handleSelectChange"
+        :item-text="(item) => t(item.label, item.label)"
+        item-value="index"
+        v-model="selectIndex"
+        v-bind="vuetifyProps('v-select')"
+        @focus="isFocused = true"
+        @blur="isFocused = false"
+      ></v-select>
+    </v-hover>
+    <dispatch-renderer
+      v-if="selectedIndex !== undefined && selectedIndex !== null"
+      :schema="indexedOneOfRenderInfos[selectedIndex].schema"
+      :uischema="indexedOneOfRenderInfos[selectedIndex].uischema"
+      :path="control.path"
+      :renderers="control.renderers"
+      :cells="control.cells"
+      :enabled="control.enabled"
+    />
 
     <v-dialog v-model="dialog" persistent max-width="600" @keydown.esc="cancel">
       <v-card>
         <v-card-title class="text-h5"> Clear form? </v-card-title>
 
         <v-card-text>
-          Your data will be cleared if you navigate away from this tab. Do you
-          want to proceed?
+          Your data will be cleared if you select this new option. Do you want
+          to proceed?
         </v-card-text>
 
         <v-card-actions>
@@ -55,13 +62,13 @@
 
 <script lang="ts">
 import {
+  CombinatorSubSchemaRenderInfo,
   ControlElement,
   createCombinatorRenderInfos,
+  createDefaultValue,
   isOneOfControl,
   JsonFormsRendererRegistryEntry,
   rankWith,
-  createDefaultValue,
-  CombinatorSubSchemaRenderInfo,
 } from '@jsonforms/core';
 import {
   DispatchRenderer,
@@ -69,26 +76,25 @@ import {
   RendererProps,
   useJsonFormsOneOfControl,
 } from '@jsonforms/vue2';
-import {
-  VDialog,
-  VCard,
-  VCardTitle,
-  VCardText,
-  VCardActions,
-  VSpacer,
-  VBtn,
-  VTabs,
-  VTab,
-  VTabsItems,
-  VTabItem,
-} from 'vuetify/lib';
-import { defineComponent, ref } from 'vue';
-import { useVuetifyControl } from '../util';
-import { CombinatorProperties } from './components';
 import isEmpty from 'lodash/isEmpty';
+import { defineComponent, ref } from 'vue';
+import {
+  VBtn,
+  VCard,
+  VCardActions,
+  VCardText,
+  VCardTitle,
+  VDialog,
+  VHover,
+  VSelect,
+  VSpacer,
+} from 'vuetify/lib';
+import { DisabledIconFocus } from '../controls/directives';
+import { useTranslator, useVuetifyControl } from '../util';
+import { CombinatorProperties } from './components';
 
 const controlRenderer = defineComponent({
-  name: 'one-of-renderer',
+  name: 'one-of-select-renderer',
   components: {
     DispatchRenderer,
     CombinatorProperties,
@@ -99,10 +105,11 @@ const controlRenderer = defineComponent({
     VCardActions,
     VSpacer,
     VBtn,
-    VTabs,
-    VTab,
-    VTabsItems,
-    VTabItem,
+    VSelect,
+    VHover,
+  },
+  directives: {
+    DisabledIconFocus,
   },
   props: {
     ...rendererProps<ControlElement>(),
@@ -111,21 +118,25 @@ const controlRenderer = defineComponent({
     const input = useJsonFormsOneOfControl(props);
     const control = (input.control as any).value as typeof input.control;
 
-    const selectedIndex = ref(control.indexOfFittingSchema || 0);
-    const tabIndex = ref(selectedIndex.value);
+    const selectedIndex = ref(control.indexOfFittingSchema);
+    const selectIndex = ref(selectedIndex.value);
     const newSelectedIndex = ref(0);
     const dialog = ref(false);
+    const t = useTranslator();
 
     return {
       ...useVuetifyControl(input),
       selectedIndex,
-      tabIndex,
+      selectIndex,
       dialog,
       newSelectedIndex,
+      t,
     };
   },
   computed: {
-    oneOfRenderInfos(): CombinatorSubSchemaRenderInfo[] {
+    indexedOneOfRenderInfos(): (CombinatorSubSchemaRenderInfo & {
+      index: number;
+    })[] {
       const result = createCombinatorRenderInfos(
         // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
         this.control.schema.oneOf!,
@@ -135,17 +146,20 @@ const controlRenderer = defineComponent({
         this.control.path,
         this.control.uischemas
       );
-      return result.filter((info) => info.uischema);
+
+      return result
+        .filter((info) => info.uischema)
+        .map((info, index) => ({ ...info, index: index }));
     },
   },
   methods: {
-    handleTabChange(): void {
+    handleSelectChange(): void {
       if (this.control.enabled && !isEmpty(this.control.data)) {
         this.dialog = true;
         this.$nextTick(() => {
-          this.newSelectedIndex = this.tabIndex;
+          this.newSelectedIndex = this.selectIndex;
           // revert the selection while the dialog is open
-          this.tabIndex = this.selectedIndex;
+          this.selectIndex = this.selectedIndex;
         });
         // this.$nextTick does not work so use setTimeout
         setTimeout(() =>
@@ -154,24 +168,28 @@ const controlRenderer = defineComponent({
         );
       } else {
         this.$nextTick(() => {
-          this.selectedIndex = this.tabIndex;
+          this.selectedIndex = this.selectIndex;
         });
       }
     },
     confirm(): void {
-      this.openNewTab();
+      this.newSelection();
       this.dialog = false;
     },
     cancel(): void {
       this.newSelectedIndex = this.selectedIndex;
       this.dialog = false;
     },
-    openNewTab(): void {
+    newSelection(): void {
       this.handleChange(
         this.path,
-        createDefaultValue(this.oneOfRenderInfos[this.newSelectedIndex].schema)
+        this.newSelectedIndex !== undefined && this.newSelectedIndex !== null
+          ? createDefaultValue(
+              this.indexedOneOfRenderInfos[this.newSelectedIndex].schema
+            )
+          : {}
       );
-      this.tabIndex = this.newSelectedIndex;
+      this.selectIndex = this.newSelectedIndex;
       this.selectedIndex = this.newSelectedIndex;
     },
   },
