+++ vue2-vuetify/src/complex/index.ts
@@ -4,7 +4,7 @@ export { default as ArrayControlRenderer } from './ArrayControlRenderer.vue';
 export { default as EnumArrayRenderer } from './EnumArrayRenderer.vue';
 export { default as ObjectRenderer } from './ObjectRenderer.vue';
 export { default as OneOfRenderer } from './OneOfRenderer.vue';
-export { default as OneOfSelectRenderer } from './OneOfSelectRenderer.vue';
+export { default as OneOfTabRenderer } from './OneOfTabRenderer.vue';
 
 import { entry as allOfRendererEntry } from './AllOfRenderer.vue';
 import { entry as anyOfRendererEntry } from './AnyOfRenderer.vue';
@@ -12,7 +12,7 @@ import { entry as arrayControlRendererEntry } from './ArrayControlRenderer.vue';
 import { entry as enumArrayRendererEntry } from './EnumArrayRenderer.vue';
 import { entry as objectRendererEntry } from './ObjectRenderer.vue';
 import { entry as oneOfRendererEntry } from './OneOfRenderer.vue';
-import { entry as oneOfSelectRendererEntry } from './OneOfSelectRenderer.vue';
+import { entry as oneOfTabRendererEntry } from './OneOfTabRenderer.vue';
 
 export const complexRenderers = [
   allOfRendererEntry,
@@ -21,5 +21,5 @@ export const complexRenderers = [
   enumArrayRendererEntry,
   objectRendererEntry,
   oneOfRendererEntry,
-  oneOfSelectRendererEntry,
+  oneOfTabRendererEntry,
 ];
